Add gost-yescrypt backend and tests.
gost-yescrypt uses the output from the yescrypt hashing method in
place of a hmac message.  Thus, the yescrypt crypto properties
are superseeded by the GOST R 34.11-2012 (Streebog) hash function
with a 256 bit digest.

This hashing method is useful in applications that need modern
password hashing methods, but require to rely on the cryptographic
properties of GOST algorithms.

The GOST R 34.11-2012 (Streebog) hash function was developed by
the Center for Information Protection and Special Communications
of the Federal Security Service of the Russian Federation with
participation of the Open joint-stock company "Information
Technologies and Communication Systems" (InfoTeCS JSC) and has
been published by the IETF as RFC 6986.

See:  https://tools.ietf.org/html/rfc6986