Convert existing manpages to BSD mdoc format.
mdoc is much more semantic than man, and is therefore a better
starting point for generation of HTML pages.  Also, the “mandoc”
program for generating HTML from mdoc source is actually maintained,
unlike “man2html.”

This conversion deliberately ignores some of the documented mdoc
conventions: in particular, I think SEE ALSO belongs at the end, and
it’s better to put the pages in the same section as the current page
first in the SEE ALSO list.

This conversion also winds up rendering slightly better with GNU
“nroff -mdoc” than with mandoc, because mandoc doesn’t implement
macros inside tbl cells; this is a known missing feature that’s
already on their to-do list.