Don't use __GNUC_PREREQ nor __GNUC_PREREQ__ (issue #40).
These are not official GCC predefined macros; they are macros defined
by GNU libc and some versions of BSD libc for internal use by their
own headers, and we shouldn't be using them without checking for their
availability first.  Since there's only one use, it's easier just not
to use them at all.

Note that in order to keep the replacement #if condition tractable, we
now don't use [static N] for GCC versions older than 5.0; formerly it
was used all the way back to 4.6.  5.0 came out in 2015, and this is
just a minor optimization, so I think we can live with that.