Replace AC_C_BIGENDIAN with a new, fat-binary friendly macro.
AC_C_BIGENDIAN tries to determine the endianness of the host
architecture at configure time.  This isn’t always possible.  For
instance, when compiling “fat” binaries for MacOS, the compiler may be
simultaneously producing (big-endian) PowerPC and (little-endian) x86
machine code, so there’s no right answer at configure time.  But there
_is_ a right answer at compile time, because the compiler’s going to
set appropriate predefined macros for each side of the fat
compilation.

The new zw_C_ENDIANNESS only figures out _what macros to use_ at
configure time; the actual decision of big- or little-endian is
deferred to when crypt-port.h is preprocessed.  There is still a
regrettably long list of macros to check, but it’s hidden away in
m4/zw_endianness.m4 and shouldn’t need to concern people most of the
time.

As a side effect, limits.h is now included by crypt-port.h and
therefore does not need to be directly included anymore.

Note that zw_C_ENDIANNESS is prepared for the possibility of a
“PDP-endian” architecture (bytes ordered 3, 4, 1, 2 within a 32-bit
word, where 1 is the lowest-order byte and 4 the highest) but
crypt-port.h will #error out if that case turns out to be true.