Drop code under #if HAVE_POSIX_MEMALIGN.
alg-yescrypt-platform.c contains code to use posix_memalign if mmap is
unavailable, but it’s not properly wired up to autoconf, so it never
actually gets used.  It is not worth attempting to use posix_memalign
when mmap is not available; the existing fallback to malloc + manual
alignment is good enough for that situation.