Restructure ifdeffage in crypt_bigcrypt_rn for analyzability.
Coverity gets confused by the combination of #if and if conditions
near the top of crypt_bigcrypt_rn, and thinks that one of the code
paths is unreachable.  Reorganize the code so that the compile-time
condition is always the same and the preprocessor condition only
changes the action taken when that condition is true.  Actual behavior
in all configurations should be unchanged.