Reduce the number of methods that can be the default for new hashes.
We don’t want to be accidentally encouraging use of gost-yescrypt,
scrypt, or the bug-compatibility bcrypt variants by people who do not
have a specific need for them.  Therefore, there is now a DEFAULT flag
in hashes.lst again, but it can appear multiple times; crypt_gensalt
uses the first enabled entry with a DEFAULT flag.

The set of hashes that can currently be used as the default for new
hashes is yescrypt ($y$), bcrypt ($2b$ only), sha512crypt ($6$) and
sha256crypt ($5$).