Make bcrypt variants independently configurable.
There are now separate INCLUDE_ macros for all four variants of
bcrypt.  --enable-hashes=bcrypt enables only the preferred $2b$
variant.  The $2x$ variant, which has a bug that makes collisions
easy to find for some passphrases, is no longer included in the STRONG
group, and the OS compatibility groups have been adjusted so that each
includes only the variants historically available on the OS we’re
being compatible with.

As a side-effect, this makes the “optional suffix on the name of the
gensalt_fn” column of hashes.lst unnecessary, and allows
simplification of gen-hashes.awk.

The bcrypt-related tests have not actually changed, but many entries
in test case arrays were rearranged to minimize the number of
additional ifdefs needed.