New test: crypt_gensalt should look only at the prefix.
For quite some time there’s been a chunk of code at the end of
test-gensalt.c labeled

    /* FIXME: This test is a little too specific.  It used to be in
       test-bcrypt.c and I'm not sure what it's meant to be testing.  */

I finally figured out what the point of this code is, which means I
can generalize it from $2a$-variant bcrypt to all hashing methods.
It’s testing that, when you supply a full setting string (or even a
full hashed passphrase) as the “prefix” argument to crypt_gensalt,
crypt_gensalt does not use any information from that string other than
the method-selection prefix.  (For instance, if you give it "$2a$05"
but a cost of 12, you should get back a string beginning "$2a$12".)

I wasn’t expecting the generalized test to detect any bugs, because in
libxcrypt, individual hashing methods’ gensalt procedures don’t get to
see the “prefix” string at all.  But it did!  Turns out the salt
generation for NTHASH fails to NUL-terminate its output properly and
we never noticed till now (mostly because nobody uses that thing, I
hope).  So that’s fixed and there’s extra code in the new test to make
sure we notice if the same mistake ever pops up again in the future.