Remove the gensalt function for the bcrypt x variant ($2x$).
This prefix was never intended for use when hashing new passphrases.
The only use case originally intended was to manually edit '$2a$' to
'$2x$' in passphrase hashes to knowingly take the risk yet enable
users to continue to log in when upgrading systems with buggy bcrypt
implementations to fixed versions.

There was never an intent to be able to generate new setting strings
with that prefix.  The original implementation of the bcrypt gensalt
function doesn't allow the use of the prefix '$2x'.
Thus libxcrypt must not, either.