Add an alias for crypt_gensalt_r.
The function was available in older versions (v3.1.1 and earlier)
of libxcrypt.  It has the same semantics and the same prototype as
the crypt_gensalt_rn function.

Adding such an alias makes porting applications written for those
earlier versions of libxcrypt much easier and requires less
conditionals to keep code compatibility for systems still using an
antique release.