Make crypt and crypt_gensalt use thread-local output buffers.
This change makes crypt and crypt_gensalt as thread-safe as they can
be without changing their interfaces.  Solaris already made this
change, and it’s being discussed by glibc (with suggestion that it
should be pushed upstream to the C and POSIX standards committees):
https://sourceware.org/ml/libc-alpha/2018-10/msg00437.html

Portable programs should still use the r-variants, though, because
this is not a guaranteed feature, it doesn’t make them not clobber
their output buffers on a second call, and the tradeoff is a sizeable
memory leak (CRYPT_GENSALT_OUTPUT_SIZE or sizeof(crypt_data)) for each
thread that calls either function.  (Thanks ever so much, C committee,
for adding thread-local variables without destructors for them.)
(I’m still investigating whether there’s some GNU extension we could
use to avoid the memory leak at least on Linux.)  (We could use
pthread_getspecific, but then we’d have to link libcrypt with
libpthread, and this doesn’t seem like enough reason to do that.)