Add the configure option --enable-unity-build.
Unity builds are a technique for cutting down build times.
The way it works is relatively straightforward.

Normally we would run the compiler three times, once for
each source file individually; in a unity build we instead
compile all these sources within a single unit.

The simplest approach is to create a new source file that
looks like this:

    #include<src1.c>
    #include<src2.c>
    #include<src3.c>
    …

This generated source file is then compiled rather than the
individual source files one by one.

The exact speedup depends on the code base, of course, but it
is not uncommon to obtain more than 50% speedup in compile times.
This happens even though the Unity build uses only one CPU
whereas individual compiles can be run in parallel.

Unity builds can also lead to faster code, because the compiler
can do more aggressive optimizations (e.g. inlining).

The downside is that incremental builds are as slow as full
rebuilds (because that is what they are).  Unity compiles also
use more memory, which may become an issue in certain scenarios.