Apply full read-only relocations to generated binaries, if supported.
The relro flag ensures that the ELF sections are reordered so that the
ELF internal data sections (.got, .dtors, etc.) precede the program's
data sections (.data and .bss) and that the non-PLT GOT is read-only.
In case of a bss or data overflow bug the relro flags protects the ELF
internal data sections from being overwritten (as the ELF sections are
reordered).

The bind-now flag tells the dynamic linker to resolve all symbols when
the program is started, or when the shared library is linked to using
dlopen, instead of deferring function call resolution to the point when
the function is first called.  Together with the relro flag the bind-now
effectively re-maps the whole GOT to be read-only.
The only downside of immediate binding causing slowed down startup of
processes as the linker has to perform all relocations at startup time,
is not that much of a problem for libxcrypt, as the only external
dependency it has is the system's libc.

The resulting full read-only relocation mitigates the well known
technique of modifying a GOT entry to get control over the program
execution flow.