Add back known-answer test for yescrypt and gost-yescrypt.
These are calculated using ctypes to call crypt_ra in the just-built
libcrypt.so.  This does not test against an independent
implementation, but it does test stability of hashes (the hash of a
passphrase today is the same as it was some time ago), which is even
more important.

There is now also a Makefile target to regenerate test-crypt-kat.inc,
since test-crypt-kat-gen.py now needs to be run from a build tree.