Convert build system from autotools to meson.
Meson (https://mesonbuild.com/) is a relatively new configuration and
build utility for C programs.  It offers several benefits for us: Its
scripting language is substantially nicer to work with than either
autoconf’s shell+m4 or makefiles.  It seems to be faster, particularly
for the configure phase.  Its shared library build logic, unlike
libtool, will let us fix issue 96 properly [note: this PR does _not_
fix issue 96].  And, perhaps most importantly, it seems to be more
actively maintained than autoconf, which we were lugging around nearly
10,000 lines worth of patches and extensions for.

Meson does have some disadvantages: most importantly, it depends on
Python >=3.6, and a tool called ‘ninja’ (https://ninja-build.org/) for
the actual building.  Because it’s under active development (unlike
autotools) we have to worry about having a suitably new version of
meson itself installed.  [I have not actually tested this PR with
anything older than 0.53, but I believe the newest features we use
come from 0.50.]  It doesn’t have a ‘distcheck’ equivalent, and it has
a number of annoying bugs that I had to work around (see comments in
the meson.build files).  All of our packagers will have to adapt
their build machinery.

Because this patch introduces a build-time requirement for
Python >=3.6 *anyway*, I also rewrote all of the build scripts in that
language, which eliminates the dependency on AWK and comes out a whole
lot nicer to read, too.  The only exceptions are scripts/travis-install,
which runs before we have a known-good Python environment, and
scripts/travis-after-success, which is so simple that it’s easier to
keep it in sh.

All of our existing build configuration options are preserved, but
they might be spelled differently now.  See meson_options.txt at top
level.  .travis.yml and scripts/travis-build may also be useful to
crib from.

This does not by itself fix the failing MacOS build, but the remaining
problems are portability bugs in the actual code rather than bugs in
the build scripts, so they should be addressed separately.