Experiment with installing meson via pip.
The meson branch needs a substantially newer version of meson than
is available from Xenial’s apt (it has 0.29, whereas I’m sure we need
at least 0.50 and the branch has only been tested with 0.53) so let’s
try installing it from pip (in a virtualenv) instead.  It also has a
pretty old version of ninja but that should be less of an issue.