Rename symbols in the static library with objcopy, not macros.
This eliminates the need to maintain a list of non-public symbols
manually in crypt-port.h.  It will also enable some simplification to
the symbol versioning macros, and possibly let us use Meson’s feature
where it generates a shared and a static library from the same object
files.

gen-crypt-hashes-h learns to emit prototypes for hash implementation
entry point functions when they’re needed internally by some other
hash.  These were previously dealt with by special case ifdeffage in
crypt-port.h; this is tidier and will be easier to extend in the
future

Potential breakage: the xcrypt_* symbols and crypt_gensalt_r are no
longer exported as default symbols from libcrypt.so.1 (and, as a
consequence, are not present at all in libcrypt.so.2).  Comments in
libcrypt.map.in said “these need to be exported as default symbols
so they are detected by AC_CHECK_FUNCS and similar” but as far as I
can see, crypt.h’s redirections should make that work regardless.

I have further plans for libcrypt.2.map.