Rename symbols in the static library with objcopy, not macros.
This eliminates the need to maintain a list of non-public symbols
manually in crypt-port.h.  It will also enable some simplification to
the symbol versioning macros, and is a step toward not needing -DPIC,
which is the primary obstacle to generating the static and the shared
library from the same object files.

gen-crypt-hashes-h learns to emit prototypes for hash implementation
entry point functions when they’re needed internally by some other
hash.  These were previously dealt with by special case ifdeffage in
crypt-port.h; this is tidier and will be easier to extend in the
future

Potential breakage: the xcrypt_* symbols and crypt_gensalt_r are no
longer exported as default symbols from libcrypt.so.1 (and, as a
consequence, are not present at all in libcrypt.so.2).  Comments in
libcrypt.map.in said “these need to be exported as default symbols
so they are detected by AC_CHECK_FUNCS and similar” but as far as I
can see, crypt.h’s redirections should make that work regardless.
(This *might* fix issue #100.)

Potential breakage: we now require a tool that can rename symbols in a
static library.  configure (sp. zw_shlib_flags.m4) will now attempt to
detect tools that are known to work for this, and will disable the
static library build if they are not found, or bomb out if
--enable-static was given on the command line.  Currently, configure
knows about two such tools: GNU binutils (nm, objcopy), and Agner
Fog’s objconv.  It is straightforward to add support for more.

In CI, GNU binutils will be used on Linux platforms, and objconv will
be used on OSX, because binutils’ support for Mach-O is known to be
somewhat buggy.  (It does appear to work for this use case, but this
way we exercise both code paths in the build machinery.)