Explicitly specify build_by_default: and install: for all targets.
It turns out that meson’s defaults for build_by_default: and install:
for executable and library targets are Wrong, and this is why we were
wasting time compiling the shared library when we only wanted the
static library, and vice versa.  For consistency, set them for
custom_target() invocations as well, even though the defaults for
those are less Wrong and were actually doing what we wanted in this case.