Replace awk and shell scripts with Python scripts.
This should be more maintainable going forward.  It also allows
simplification of the Makefile and configure.ac logic in a few places.

The new Python scripts are without a file extension, as is recommended
by Python coding convention, *except* that the test scripts have a .py
extension so we can use PY_LOG_COMPILER to tell automake which
interpreter to use for them.

The Travis scripts also move into scripts/.  .travis_script.sh is
rewritten in Python, but the other two remain shell scripts, because
travis-install runs before we have a usable Python environment, and
travis-after-success is too simple to be worth rewriting in Python.

skip-if-exec-format-error also remains a shell script, in this case
not because the rewrite isn’t worth it (any shell script that needs
temporary files is worth rewriting in not-shell just because of that),
but because it’s only used in cross-compilation so the rewrite would
be too hard to *test*.

The build logic executed by travis-build is revised somewhat: most
importantly, we now explicitly pass --enable-shared to configure
in all configurations that aren’t passing --disable-shared.  This
causes it to bomb out when shared libraries aren’t possible, instead
of silently disabling them and going on.  We also abandon building
fat binaries on OSX, as this only works with Xcode clang, and Apple
seems to be phasing out support for the 32-bit ABI anyway (see e.g.
http://netkas.org/?p=1491 .)