Replace libtool with custom autoconf and automake logic.
This isn't exactly straightforward, and it definitely supports fewer
OSes than we used to.  Anything ELFy should still work with GCC(-alikes),
OSX is work in progress, nothing else is implemented; reasonably
modern OSes should be no problem to add back, as requested.

But the results are nice: the generated configure script shrinks by
7000 lines, the 11,000-line ltmain.sh is completely dropped from the
tarball, and the single-threaded build time goes from 18 to 12 seconds
on the same hardware.  (-j4 build time goes from 7.5 to 5.7 seconds.
CPU time goes from 15 to 10 seconds *either way*, so the improvement
is all elimination of unnecessary work, no waiting.)  (These numbers
include the time to compile all of the test programs, but not run them.)

Also, this will enable us to rename private symbols in the static
library using objcopy instead of preprocessor macros, which will be
completely automatic and less fragile.  With a little more work it may
also allow us to use the same object files for libcrypt.a and libcrypt.so
(a la Meson's b_staticpic mode).