Add static-only builds on macOS.
Also, do all static builds and all macOS builds with
--disable-obsolete-api.  It’s implied by --disable-shared, and on
macOS there is no system libcrypt.so.1 to be compatible with; in both
cases the configure script was silently turning the obsolete API off.
Explicitly saying so in .travis.yml is clearer.

Also, put “Linux” in the name: field of all the Linux builds to make
it a little easier to wade through the very long list of build
configurations.