Salt string compatibility with generic implementations.
All other existing implementations of the md5crypt, sha256crypt,
and sha512crypt hashing methods allow any ASCII character to be
present in the salt string.

We changed our implementation of these hash methods to be compliant
with the general behaviour, except for the colon ':' character,
which serves a special purpose in the Unix shadow file.

Also adapt test/badsalt.c to expect the colon characters is an invalid
salt character for all hashing methods.

Fixes #105