test/special-char-salt: New test.
Simple test to ensure compatability of our implementations of the
md5crypt, sha256crypt, and sha512crypt hashing method with the ones
found in other commonly used libcrypt libraries.  Their implementations
allow for any printable ASCII character (with one exception to the
colon ':' character, that has special meaning in the Unix shadow file)
to be present in a salt string.