Rewrite build scripts in Perl.
This introduces a new build-time requirement for Perl >=5.14.0, which
was the version shipped with Ubuntu 14.04 (Trusty).

The scripts whose language changes are:

 - build-aux/compute-symver-floor (formerly inlined in configure.ac)
 - build-aux/expand-selected-hashes
 - build-aux/gen-crypt-h
 - build-aux/gen-crypt-hashes-h
 - build-aux/gen-crypt-symbol-vers-h
 - build-aux/gen-libcrypt-map

A few other one-off uses of awk are also eliminated.

Using a more powerful scripting language allows us to unify the
parsers for hashes.conf and libcrypt.map.in, and simplify two
other aspects of the build:

 - the four files libcrypt.minver.{hurd,kfreebsd,linux,no} are
   combined into a single file libcrypt.minver
 - crypt.h is now generated in a single step from crypt.h.in, and
   does not depend on crypt-hashes.h

Error checking is also substantially more robust.