test/symbols-compat.sh: Don’t rely on -lcrypt to find libcrypt.so.1.
Some Linux distributions are now shipping libcrypt.so as a symlink to
libcrypt.so.2, but still providing libcrypt.so.1 as a compatibility
library.  The symbols-compat.sh test is still relevant on these
systems, but linking a test program against -lcrypt will not find
libcrypt.so.1.  Instead use gcc -print-file-name=libcrypt.so.1, which
will find that library regardless of what other libcrypt* files are
present.

This also has the nice fringe benefit that we don’t have to parse ldd
output anymore.

This does *not* fix the symbols-compat failures in CI, but it should
not make them worse.