test/symbols-compat.sh: compensate for a bug in nm
Some versions of GNU nm have a bug where
`nm --dynamic --with-symbol-versions`
prints the symbol version twice.
This does not invalidate the test,
but it does make the test logs difficult to read.
Postprocess out the extra copy.
(Not all versions of awk support back-references in regexes,
so it is easiest to do this by converting the postprocess logic
to a sed script.)

Still doesn’t fix the CI regressions.