test/symbols-compat.sh: Add a filter for expected differences.
The symbols-compat test is currently failing in CI because it does not
account for _intentional_ differences between older, system-provided
libcrypt.so.1 and what we currently produce:

  * All of the symbols we define with GLIBC_2.xx version tags are
    compatibility symbols (nm prints only one @); naturally,
    glibc-provided libcrypt.so.1 defines some of those symbols as
    linkable symbols (two @).

  * Older versions of libcrypt defined five symbols as linkable,
    with the XCRYPT_2.0 version tag, which are now compatibility-only:
    crypt_gensalt_r, xcrypt, xcrypt_gensalt, xcrypt_gensalt_r, and
    xcrypt_r.

Prior to commit 7a1df1bb8599ea958c7fdd30ea3c7f00487b8e08, this problem
was masked by a typo in symbols-compat.sh, which made the test not
test anything meaningful.

Also changes from logging the full contents of both the “our symbols”
and “their symbols” files when the test fails, to logging a diff
between the two, which is easier to read.