CI: Don’t create a python virtualenv.
The only thing we’re using python for, in CI, is the codecov utility,
and there is no need to create a virtualenv just for that.
This should shave a few seconds off each build.

Moreover, we can skip installing codecov in coverity and distcheck
builds, since we won’t be running it.