CI: Log identities of build and coverage tools.
In .travis_before_script.sh and .travis_after_success.sh, log the full
path and the --version output of each basic tool that we require.

The before script does the build tools, and the after script does the
coverage tools.  (This is because only the after script knows exactly
which version of gcov we’re going to run.)