Use codecov.io’s “bash uploader” instead of codecov from PyPI.
This works kind-of the same way as the Coverity driver: we retrieve a
shell script directly from codecov.io and execute it.  This is how
codecov.io currently recommends one upload coverage reports, and means
we do not need to worry about whether ‘pip’ works.  In fact, we no
longer need an install script at all.

pip is removed from the cache set, and the after-success script is
moved to the script: tag so that coverage upload failures are flagged
in the Travis dashboard.