Split up crypt-common.c, normalize naming of utility functions.
crypt-common.c holds one piece of static data and three utility
functions, that don’t have anything to do with each other.  Split it
up into one file per item.  This allows us to be more precise about
which unit tests need which files, and also paves the way for a more
precise and LTO-friendly implementation of _crypt_secure_memset.

At the same time, rename _crypt_secure_memset to explicit_bzero and
_crypt_strcpy_or_abort to strcpy_or_abort.  Both are given the _crypt_
prefix back by #defines in crypt-port.h as usual.  This is in
preparation for adding the _crypt_ prefix with objcopy magic instead
of with #defines.

strcpy_or_abort was using XCRYPT_SECURE_MEMSET, but does not need to
do so, as the data in its destination buffer is not being thrown away;
switch to memset instead.