Improve fallback implementation of explicit_bzero.
Instead of writes through a volatile pointer, which forces inefficient
code generation and _isn’t_ guaranteed to do what we want in this
context, use a regular call to memset followed by a no-op assembly
insert that is declared to read the memory that we wrote.  Also mark
the function as not inlinable, which has no effect now (it’s in a TU
by itself anyway) but may become important if we ever turn on
link-time optimization.

Also, and perhaps more important, import glibc’s test of
explicit_bzero and apply it to whatever “memory clear that can’t be
optimized out” function we have found.