Reorganize library API definitions into one file per entry point.
The new files are named e-<function-name>.c.  This makes it so users
of the static library only pull in the functions they actually use,
which is a fairly marginal benefit for this library, since almost any
use winds up pulling all of the hashing methods anyway.  More
importantly, this will facilitate replacement of the namespace-
cleanliness #defines with object file rewriting, since there is now a
sharp distinction between files that define entry points (and nothing
else) and files that don’t.

The code that dispatches operations to specific hashing methods is now
also isolated in its own file (hm-dispatch.c).  This is preparation
for revising how _that_ works pretty drastically as well.