configure.ac: export CC before invoking compute-symver-floor.
compute-symver-floor may need to run the compiler, and due to an
oversight in configure.ac it was always using ‘cc’ rather than the
compiler in $CC.  Particularly when cross compiling, it is essential
to use the same compiler here that the rest of the build would.

Issue originally reported as Debian bug #985372.