configure.ac: error out if compute-symver-floor returns nothing.
This _should_ only ever happen if there’s some sort of low-level
failure.  These are already being logged in config.log, but we
shouldn’t allow configure to succeed when they happen.  We have
a report of “sporadic” build failures due to this oversight
(issue #123).  This patch will only move the failure from the make
step to the configure step, but hopefully that will make debugging
easier.