Don’t build xcrypt.h if we’re not going to install it.
All of the functions declared by xcrypt.h are defined using
strong_alias(), which does not need a prototype to be visible.
In fact, the entire contents of xcrypt.h are already #ifdefed out
when building libxcrypt itself.  So it’s not necessary to include it
internally. If we stop including it internally, then we don’t need
to build it when it’s not going to be installed, and we can remove
the #ifndef IN_LIBCRYPT from the installed header.