Don’t apply --enable-obsolete-api-enosys mode to fcrypt.
fcrypt is the only function that has two alternative implementations,
selected by configure options, one of which is an alias for another
function.  This is interfering with my plans to handle symbol
versioning with objcopy instead of a maze of #ifdefs.

fcrypt is not part of POSIX.  In GNU libc, our most important
compatibility target, it has always been another name for crypt:
in particular it has always supported all of the same hashing
methods as crypt itself.  Therefore, it seems to me we should
treat it the same as any other obsolete alternative name for
crypt (e.g. xcrypt): either not available at all, or available
to old binaries with all the same functionality it’s ever had.

Accordingly, remove the --enable-obsolete-api-enosys version of
fcrypt.  Also, don’t bother testing it in ka-tester, since we know by
construction it’s just another name for crypt.