Avoid rebuilding the world every time the Makefile changes.
Almost everything depends on the generated headers; the generated
headers depend on the Makefile; *most* changes to the Makefile won't
affect the contents of the generated headers.  Arrange for the
timestamps of the generated headers not to change if their contents
haven't, using an idiom borrowed from GCC's Makefile.

This exposed a bug in BuildCommon.pm that caused the contents of
crypt-symbol-vers.h to change every time it was regenerated (due to
Perl’s hash randomization).  Fix that too.