lib/crypt.c: Indicate weak hashes by crypt_checksalt().
The hashes.conf file used to define the hashing methods supported
by libxcrypt defines a group of STRONG hashing methods.  Those methods,
that are not members of this group, are considered to be too weak for
still being used for new passphrases.

The crypt_checksalt() function now returns `CRYPT_SALT_METHOD_LEGACY`
for such settings.