lib/crypt.c: Fix return value of crypt_checksalt() for empty settings.
In the previous commits we implemented the return value for indicating
the use of weak hashing methonds in crypt_checksalt().  It turned out,
the implementation has a bug, that that erroneously reported settings,
which are representing an empty passphrase or an uncomputed setting for
descrypt without any salt characters to be weak instead of invalid.