Github Actions: Add workflow to run ‘make distcheck’.
This test is run only with GCC, as on Travis.

Use the ‘setup-perl’ action (shogo82148/actions-setup-perl@v1) to run
the distcheck build with Perl 5.14, thus consolidating the distcheck
build with the “oldest supported perl” build that we had on Travis.

Read the set of critic-related CPAN distributions that are expected to
be installed from .perlcriticrc and install them with cpanm.  (Some
but not all of them are available via apt, but this way we can pin
exact versions.)  Download CPAN packages over HTTPS; this might mean
we’re hitting one overloaded server on someone’s desk somewhere, but
CPAN packages don’t necessarily have any built-in cryptographic
integrity verification, and even when they do you have to manually set
up PGP keys for cpanm --verify to work.  Transport security is at
least *something*.