Github Actions: Add workflow for code coverage monitoring.
This is the second of two patches being landed on trunk ahead of the
complete migration from Travis CI to Github Actions; see the commit
message for 407a546a1a8477dd18d396cb6e0533032062381e for more
explanation.

This workflow may not yet achieve code coverage matching what we had
with Travis.  We are only building on Linux for now, and only with a
handful of configure-time variations.