Overhaul the badsalt test.
Test patterns are now mostly generated rather than manually coded into
a big table.  Not reading past the end of the “setting” part of the
string is tested more thoroughly (this would have caught the sunmd5 $$
bug if it had been available at the time).  Test logs are tidier.

This change exposed a lacuna in handling of text beyond the end of
the setting: we were not rejecting e.g. ‘$3$__not_used_:c809a450df09a3’
because nthash ignores everything after ‘$3$’ and nothing else detects
the colon.  Now setting strings containing ‘:’ or a newline are
rejected in generic code.  (I would like to expand this to match the
rule laid out in crypt(5),

    Hashed passphrases are always entirely printable ASCII, and do not
    contain any whitespace or the characters ‘:’, ‘;’, ‘*’, ‘!’, or
    ‘\’.  (These characters are used as delimiters and special markers
    in the passwd(5) and shadow(5) files.)

but I am worried that that will trip up people with stored passphrases
with weird salt, like in <https://github.com/besser82/libxcrypt/issues/105>,
so I’m not doing it yet.