Disable Travis CI.
We are migrating from Travis CI to Github Actions because the Travis build
farm has become too unreliable.  As of spring 2021 any given build was
likely to have at least one job fail due to problems within the build
infrastructure (e.g. timeouts downloading dependencies).  This was at the
same time the company was failing to communicate honestly about the
withdrawal of their free-to-FLOSS-projects tier of service: see for example

 - https://blog.travis-ci.com/2021-05-07-orgshutdown
 - https://travis-ci.community/t/org-com-migration-unexpectedly-comes-with-a-plan-change-for-oss-what-exactly-is-the-new-deal/10567
 - https://travis-ci.community/t/extremely-poor-official-communication-of-the-org-shutdown/10568

We appreciate that “free to FLOSS projects” is becoming untenable for many
organizations due to abuse: see for example

 - https://github.blog/2021-04-22-github-actions-update-helping-maintainers-combat-bad-actors/
 - https://man.sr.ht/ops/builds.sr.ht-migration.md
 - https://about.gitlab.com/blog/2021/05/17/prevent-crypto-mining-abuse/

where three competing CI providers (including Github) announce,
without mincing any words, that their free CI service is being abused
to mine cryptocurrency and, as a countermeasure, they are requiring at
least the _ability_ to charge all users for their CPU time costs.
If Travis had been this transparent we might have been willing to work
with their new model, but they weren’t.

The Travis configuration and build scripts are only disabled, not removed,
with this commit; they will be removed when the migration is complete.
Our original plan for this migration was to do all the work in a topic
branch (see pull request #131) and land it only when complete, but due to
major gaps in our code coverage logs on trunk (in turn because of Travis’
unreliability) it has proved impossible to tell whether changes on the
branch are having the desired effects.  We are therefore making a clean
break on trunk, with this commit and the next one, ahead of the rest of the
changes.