List policies exhaustively in .perlcriticrc and enforce this.
.perlcriticrc now explicitly mentions every single policy that is
available from the policy packages we use, as either enabled or
disabled.  Furthermore .perlcriticrc lists, in comments, all of the
expected policy sets and related distributions that should be
available, with version pins; this is now the canonical home for this
information.

A new ancillary script, build-aux/check-perlcritic-config, validates
that the list of policies in .perlcriticrc is 1:1 with the set of
available policies.  It is run by ‘make distcheck.’  (Todo: convert to
a local perlcritic policy so that every run of perlcritic does this
validation.)

These changes will hopefully prevent future problems along the lines
of 5ad98a9dca84f064e7dc19049afbc478d7320f97 (reverted by
55880f9d2e057b83ccaf29132f01a16d7285a7a7), where two developers did
not have the same set of perlcritic policies installed and therefore
disagreed on whether a ‘no critic’ annotation was necessary.  They
will also ensure that the set of perlcritic policies being tested in
CI is exactly what we expect.

The downside of this change is that .perlcriticrc is now harder to
read through and find all the policies with configuration.  Better
ideas for organizing it are welcomed.