configure: Restore ucontext api functionality check.
In c3f01c72b303cbbb0cc8983120677edee2f3fa4b the use of the ucontext api
in the main program was removed, and with it the configure check for it.
However, the ucontext api is still used in the "explicit_bzero" test and
this has the same problem as described in the comment:

> The ucontext.h functions that we use were withdrawn from
> POSIX.1-2008, so the existence of the header does not prove
> we can use the functions.

Thus, restore the full configure check and use it instead of the header
check for the explicit_bzero test.

See https://bugs.gentoo.org/838172

Signed-off-by: Björn Esser <besser82@fedoraproject.org>