configure: Only test the makecontext signature we need
The test/explicit-bzero.c test uses a start routine without any
arguments.  There is no need for the multi-argument version.

This avoids a build failure with glibc and future compilers.
The GNU C library declares the makecontext callback of
type void (*) (void), so no cast is needed.  On other systems,
the type may be the (currently distinct) type void (*) (),
but given that this only affects the ability to execute a test,
no further machinery is added here to detect that different type.