Backport new-style 'PKG_PROG_PKG_CONFIG' macro from pkg-config 0.29.2.
Since the definition of this macro has been changed to error-out if a new
introduced second parameter is ommited and the pkg-conf executable is not
found on the system, the configure script from the release tarballs will
fail in certain minimal bootstrap environments, that do not include a
pkg-config executable (or the provided version is too old).

The backport of a compatibility is needed to be able to regenerate the
build-system on older systems, which do not offer a variant of the macro
definition with an optional second parameter.

Fixes #198.